DXP-Decrypt/requirements.txt
python3-pip
azure-storage-blob
requests
python-decouple
subprocess


